internal: types adjusts
