+++ generators/angular/types.d.ts
@@ -16,8 +16,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { ApplicationType } from '../../lib/types/application/application.js';
 import type { Entity } from '../../lib/types/application/index.js';
+import type { ApplicationAll } from '../base-application/types-all.js';
 
 export type AngularEntity = Entity & {
   /**
@@ -35,4 +35,4 @@ export type AngularApplication = {
   /** @experimental to be replaced with needles */
   angularEntities?: AngularEntity[];
   angularLocaleId: string;
-} & ApplicationType<AngularEntity>;
+} & ApplicationAll<AngularEntity>;
