+++ generators/base-application/types-all.d.ts
@@ -1,1 +1,147 @@
+/* eslint-disable @typescript-eslint/consistent-type-imports */
+import type { ExportApplicationPropertiesFromCommand } from '../../lib/command/types.js';
+import { Entity } from '../../lib/types/application/entity.js';
+import type { ClientApplication, ClientSourceType } from '../client/types.js';
+import { DockerSourceType } from '../docker/types.js';
+import { LanguagesSource } from '../languages/types.js';
+import type { SpringBootApplication, SpringBootSourceType } from '../server/types.js';
+import type { OptionWithDerivedProperties } from './application-options.js';
+import type { Application as BaseApplication, Source as BaseApplicationSource } from './types.js';
+
+/* ApplicationType Start */
+type MicroservicesArchitectureApplication = {
+  microfrontend: boolean;
+  gatewayServerPort: number;
+};
+
+type GatewayApplication = MicroservicesArchitectureApplication & {
+  microfrontends: string[];
+};
+
+/*
+Deterministic option causes types to be too complex
+type ApplicationType = DeterministicOptionWithDerivedProperties<
+  'applicationType',
+  ['monolith', 'gateway', 'microservice'],
+  [Record<string, never>, GatewayApplication, MicroservicesArchitectureApplication]
+>;
+*/
+type ApplicationProperties = OptionWithDerivedProperties<'applicationType', ['monolith', 'gateway', 'microservice']> &
+  GatewayApplication &
+  MicroservicesArchitectureApplication;
+
+/* ApplicationType End */
+
+/* AuthenticationType Start */
+/*
+Deterministic option causes types to be too complex
+type UserManagement =
+  | {
+      skipUserManagement: true;
+      generateUserManagement: false;
+      generateBuiltInUserEntity?: false;
+      generateBuiltInAuthorityEntity: false;
+    }
+  | {
+      skipUserManagement: false;
+      generateBuiltInUserEntity?: boolean;
+      generateUserManagement: true;
+      user: any;
+      userManagement: any;
+      generateBuiltInAuthorityEntity: boolean;
+      authority: any;
+    };
+    */
+type UserManagement<Entity> = {
+  skipUserManagement: boolean;
+  generateUserManagement: boolean;
+  generateBuiltInUserEntity?: boolean;
+  generateBuiltInAuthorityEntity: boolean;
+  user: Entity;
+  userManagement: Entity;
+  authority: Entity;
+};
+
+type JwtApplication = {
+  jwtSecretKey: string;
+};
+
+type Oauth2Application = {
+  syncUserWithIdp?: boolean;
+};
+
+type SessionApplication = {
+  rememberMeKey: string;
+};
+
+/*
+Deterministic option causes types to be too complex
+type AuthenticationType = DeterministicOptionWithDerivedProperties<
+  'authenticationType',
+  ['jwt', 'oauth2', 'session'],
+  [JwtApplication, Oauth2Application, SessionApplication]
+>;
+*/
+type AuthenticationProperties<Entity> = OptionWithDerivedProperties<'authenticationType', ['jwt', 'oauth2', 'session']> &
+  UserManagement<Entity> &
+  JwtApplication &
+  Oauth2Application &
+  SessionApplication;
+
+/* AuthenticationType End */
+
+type QuirksApplication = {
+  cypressBootstrapEntities?: boolean;
+};
+
+export type ApplicationAll<E = Entity> = BaseApplication &
+  QuirksApplication &
+  AuthenticationProperties<E> &
+  SpringBootApplication &
+  ClientApplication &
+  ExportApplicationPropertiesFromCommand<typeof import('../docker/command.ts').default> &
+  ExportApplicationPropertiesFromCommand<typeof import('../git/command.ts').default> &
+  ExportApplicationPropertiesFromCommand<typeof import('../gradle/command.ts').default> &
+  ExportApplicationPropertiesFromCommand<typeof import('../project-name/command.ts').default> &
+  ExportApplicationPropertiesFromCommand<typeof import('../spring-boot/command.ts').default> &
+  import('../docker/types.js').DockerApplicationType &
+  ApplicationProperties & {
+    clientRootDir: string;
+    clientSrcDir: string;
+    clientTestDir?: string;
+    clientDistDir?: string;
+    devServerPort: number;
+    pages: string[];
+
+    serverPort: number;
+    backendType?: string;
+    backendTypeJavaAny?: boolean;
+    backendTypeSpringBoot?: boolean;
+    temporaryDir?: string;
+
+    hipsterName?: string;
+    hipsterProductName?: string;
+    hipsterHomePageProductName?: string;
+    hipsterStackOverflowProductName?: string;
+    hipsterBugTrackerProductName?: string;
+    hipsterChatProductName?: string;
+    hipsterTwitterUsername?: string;
+    hipsterDocumentationLink?: string;
+    hipsterTwitterLink?: string;
+    hipsterProjectLink?: string;
+    hipsterStackoverflowLink?: string;
+    hipsterBugTrackerLink?: string;
+    hipsterChatLink?: string;
+
+    dockerServicesDir?: string;
+    dockerServices?: string[];
+    prettierFolders?: string;
+    prettierExtensions?: string;
+
+    loginRegex?: string;
+    jsLoginRegex?: string;
+  };
+
 export { Options as AllOptions, Config as AllConfig } from './types.js';
+
+export type SourceAll = BaseApplicationSource & SpringBootSourceType & ClientSourceType & LanguagesSource & DockerSourceType;
