+++ generators/base-workspaces/support/config.ts
@@ -1,7 +1,7 @@
 import { mutateData, pickFields } from '../../../lib/utils/index.ts';
 
 import { monitoringTypes, serviceDiscoveryTypes } from '../../../lib/jhipster/index.ts';
-import type { PlatformApplication } from '../../base-application/types.js';
+import type { WorkspacesApplication } from '../types.js';
 
 const { PROMETHEUS, ELK } = monitoringTypes;
 const { NO: NO_SERVICE_DISCOVERY, CONSUL, EUREKA } = serviceDiscoveryTypes;
@@ -10,7 +10,7 @@ const { NO: NO_SERVICE_DISCOVERY, CONSUL, EUREKA } = serviceDiscoveryTypes;
  * @param {Object} config - config to load config from
  * @param {import('./base-application/types.js').PlatformApplication} dest - destination context to use default is context
  */
-export const loadPlatformConfig = ({ config, application }: { config: any; application: PlatformApplication }) => {
+export const loadPlatformConfig = ({ config, application }: { config: any; application: WorkspacesApplication }) => {
   mutateData(application, pickFields(config, ['serviceDiscoveryType', 'monitoring']));
 };
 
@@ -27,7 +27,7 @@ export const loadDerivedServerAndPlatformProperties = ({ application }: { applic
  * @param {import('./bootstrap-application-server/types').SpringBootApplication} dest - destination context to use default is context
  * @param {import('./base-application/types.js').PlatformApplication} dest - destination context to use default is context
  */
-export const loadDerivedPlatformConfig = ({ application }: { application: PlatformApplication }) => {
+export const loadDerivedPlatformConfig = ({ application }: { application: WorkspacesApplication }) => {
   application.monitoringElk = application.monitoring === ELK;
   application.monitoringPrometheus = application.monitoring === PROMETHEUS;
   loadDerivedServerAndPlatformProperties({ application });
