+++ generators/base-workspaces/types.d.ts
@@ -16,6 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import type { OptionWithDerivedProperties } from '../base-application/application-options.js';
 import type { Config as BaseConfig, Features as BaseFeatures, Options as BaseOptions } from '../base/index.js';
 
 export type { Source } from '../base/types.js';
@@ -32,3 +33,9 @@ export type Config = BaseConfig & {
 export type Features = BaseFeatures;
 
 export type Options = BaseOptions;
+
+type ServiceDiscoveryApplication = OptionWithDerivedProperties<'serviceDiscoveryType', ['no', 'eureka', 'consul']>;
+
+type MonitoringApplication = OptionWithDerivedProperties<'monitoring', ['no', 'elk', 'prometheus']>;
+
+export type WorkspacesApplication = ServiceDiscoveryApplication & MonitoringApplication;
